refactor: Move encryption and key exchange files under stream/encryption.
