+++ src/index.ts
@@ -10,7 +10,7 @@ import { StreamrClient } from './StreamrClient'
 export * from './StreamrClient'
 export * from './Config'
 export * from './stream'
-export * from './stream/Encryption'
+export * from './stream/encryption/Encryption'
 export * from './stream/StreamPart'
 export * from './stream/StorageNode'
 export * from './subscribe'
