+++ src/stream/encryption/BrowserStore.ts
