+++ src/stream/encryption/KeyExchange.ts
@@ -4,14 +4,14 @@ import {
 import mem from 'mem'
 import pMemoize from 'p-memoize'
 
-import { uuid, Defer } from '../utils'
-import Scaffold from '../utils/Scaffold'
+import { uuid, Defer } from '../../utils'
+import Scaffold from '../../utils/Scaffold'
 
-import { validateOptions } from './utils'
+import { validateOptions } from '../utils'
 import EncryptionUtil, { GroupKey, GroupKeyish, StreamMessageProcessingError } from './Encryption'
-import type { Subscription } from '../subscribe'
-import { StreamrClient } from '../StreamrClient'
-import PersistentStore from './PersistentStore'
+import type { Subscription } from '../../subscribe'
+import { StreamrClient } from '../../StreamrClient'
+import GroupKeyStore from './GroupKeyStore'
 
 const KEY_EXCHANGE_STREAM_PREFIX = 'SYSTEM/keyexchange'
 
@@ -62,130 +62,6 @@ function getKeyExchangeStreamId(address: Address) {
         return address // prevent ever double-handling
     }
     return `${KEY_EXCHANGE_STREAM_PREFIX}/${address.toLowerCase()}`
-}
-
-type GroupKeyStoreOptions = {
-    clientId: string,
-    streamId: string,
-    groupKeys: [GroupKeyId, GroupKey][]
-}
-
-export class GroupKeyStore {
-    store
-    currentGroupKeyId: GroupKeyId | undefined // current key id if any
-    nextGroupKeys: GroupKey[] = [] // the keys to use next, disappears if not actually used. Max queue size 2
-
-    constructor({ clientId, streamId, groupKeys }: GroupKeyStoreOptions) {
-        this.store = new PersistentStore({ clientId, streamId })
-
-        groupKeys.forEach(([groupKeyId, groupKey]) => {
-            GroupKey.validate(groupKey)
-            if (groupKeyId !== groupKey.id) {
-                throw new Error(`Ids must match: groupKey.id: ${groupKey.id}, groupKeyId: ${groupKeyId}`)
-            }
-            // use last init key as current
-            this.currentGroupKeyId = groupKey.id
-        })
-    }
-
-    private async storeKey(groupKey: GroupKey) {
-        GroupKey.validate(groupKey)
-        const existingKey = await this.store.get(groupKey.id)
-        if (existingKey) {
-            if (!existingKey.equals(groupKey)) {
-                throw new GroupKey.InvalidGroupKeyError(
-                    `Trying to add groupKey ${groupKey.id} but key exists & is not equivalent to new GroupKey: ${groupKey}.`,
-                    groupKey
-                )
-            }
-
-            await this.store.set(groupKey.id, existingKey)
-            return existingKey
-        }
-
-        await this.store.set(groupKey.id, groupKey)
-        return groupKey
-    }
-
-    async has(id: GroupKeyId) {
-        if (this.currentGroupKeyId === id) { return true }
-
-        if (this.nextGroupKeys.some((nextKey) => nextKey.id === id)) { return true }
-
-        return this.store.has(id)
-    }
-
-    async isEmpty() {
-        return !this.nextGroupKeys.length && await this.store.size() === 0
-    }
-
-    async useGroupKey(): Promise<[GroupKey | undefined, GroupKey | undefined]> {
-        const nextGroupKey = this.nextGroupKeys.pop()
-        // First use of group key on this stream, no current key. Make next key current.
-        if (!this.currentGroupKeyId && nextGroupKey) {
-            this.currentGroupKeyId = nextGroupKey.id
-            return [
-                await this.get(this.currentGroupKeyId),
-                undefined,
-            ]
-        }
-
-        // Keep using current key (empty next)
-        if (this.currentGroupKeyId != null && !nextGroupKey) {
-            return [
-                await this.get(this.currentGroupKeyId),
-                undefined
-            ]
-        }
-
-        // Key changed (non-empty next). return current + next. Make next key current.
-        if (this.currentGroupKeyId != null && nextGroupKey != null) {
-            const prevId = this.currentGroupKeyId
-            this.currentGroupKeyId = nextGroupKey.id
-            const prevGroupKey = await this.get(prevId)
-            // use current key one more time
-            return [
-                prevGroupKey,
-                nextGroupKey,
-            ]
-        }
-
-        // Generate & use new key if none already set.
-        await this.rotateGroupKey()
-        return this.useGroupKey()
-    }
-
-    async get(id: GroupKeyId) {
-        return this.store.get(id)
-    }
-
-    async clear() {
-        this.currentGroupKeyId = undefined
-        this.nextGroupKeys.length = 0
-        return this.store.clear()
-    }
-
-    async rotateGroupKey() {
-        return this.setNextGroupKey(GroupKey.generate())
-    }
-
-    async add(groupKey: GroupKey) {
-        return this.storeKey(groupKey)
-    }
-
-    async setNextGroupKey(newKey: GroupKey) {
-        GroupKey.validate(newKey)
-        this.nextGroupKeys.unshift(newKey)
-        this.nextGroupKeys.length = Math.min(this.nextGroupKeys.length, 2)
-        await this.storeKey(newKey)
-    }
-
-    async rekey() {
-        const newKey = GroupKey.generate()
-        await this.storeKey(newKey)
-        this.currentGroupKeyId = newKey.id
-        this.nextGroupKeys.length = 0
-    }
 }
 
 type GroupKeysSerialized = Record<GroupKeyId, GroupKeyish>
@@ -663,6 +539,7 @@ export function SubscriberKeyExchange(client: StreamrClient, { groupKeys = {} }:
             if (!streamMessage.newGroupKey) { return }
             const streamId = streamMessage.getStreamId()
             const groupKeyStore = await getGroupKeyStore(streamId)
+            // newGroupKey has been converted into GroupKey
             const newGroupKey: unknown = streamMessage.newGroupKey
             await groupKeyStore.add(newGroupKey as GroupKey)
         },
